[SPARK][PERF] Step 7: Use Platform.getShort in putShortsFromIntsLittleEndian
In OnHeapColumnVector and OffHeapColumnVector, replace
(short) Platform.getInt(src, srcOffset) with Platform.getShort(src, srcOffset)
in the LE branch of putShortsFromIntsLittleEndian.

On LE platform, the low 2 bytes of each 4-byte INT32 contain the short
value. Reading a short directly reads only 2 bytes instead of 4.

JMH results on Apple M3 (ops/s, higher is better):
- HEAP/ON_HEAP:    1,642,031 -> 1,667,871  (+2%)
- HEAP/OFF_HEAP:   1,679,673 -> 1,700,241  (+1%)
- DIRECT/ON_HEAP:  1,678,874 -> 1,347,958  (-20%)
- DIRECT/OFF_HEAP: 1,379,053 -> 1,354,159  (~same)

Neutral on M3 - the JIT likely already optimizes the int-to-short
truncation. Will validate on x86.

Also adds final summary of all optimization steps to benchmark results
document.

+++ sql/core/src/main/java/org/apache/spark/sql/execution/vectorized/OffHeapColumnVector.java
@@ -277,8 +277,10 @@ public final class OffHeapColumnVector extends WritableColumnVector {
           (short) Integer.reverseBytes(Platform.getInt(src, srcOffset)));
       }
     } else {
+      // On LE platform, the low 2 bytes of each 4-byte INT32 contain the short value.
+      // Reading a short directly avoids reading 4 bytes and truncating.
       for (int i = 0; i < count; ++i, srcOffset += 4, dstOffset += 2) {
-        Platform.putShort(null, dstOffset, (short) Platform.getInt(src, srcOffset));
+        Platform.putShort(null, dstOffset, Platform.getShort(src, srcOffset));
       }
     }
   }
